step2 중간 커밋. static method in interface 문제. 2020/05/15 09:28
