+++ src/main/java/step2/Calculator.java
@@ -1,0 +1,7 @@
+package step2;
+
+public interface Calculator {
+
+    static int calculate(int result, int nextNumber, String operator);
+
+}
